[Mix] Implement llvm.mix intrinsic in terms of library API calls
This is the initial commit implementing basic specializer generation.
The Mix pass is now responsible for expanding calls to `llvm.mix' and
creating custom specializer functions. No binding-time separation is
currently performed.

Specializer code generation is handled by StagedIRBuilder class that is
used to construct sequences of LLVM API calls to recreate a given piece
of IR.

LibFFI is required for running Mix tests on LLI.