[Mix] Use opaque LLVM Core types in generated code
This commit makes StagedIRBuilder use pointers to opaque structs from
LLVM Core API instead of generic `i8*' pointer type. This clarifies
generated code, cuts down some potential bitcasts, and can hopefully
help catch a bug or two in the future. The context parameter of the
specializer is now also typed as a pointer to opaque struct - this
pushes the bitcasts down to the original call site, where they can be
eliminated completely if the original context pointer is of the same
type as we require (i.e. the code uses LLVM C API).