[Mix] Delay building phi incoming lists until the values are ready
This commit removes forced staging when building phi incoming lists
since this can easily lead to indefinite cycling. Instead, a callback
interface is introduced which allows the implementation to schedule a
callback for when a value is normally staged, which is used to add a
necessary delay between constructing a phi and building its incoming
list.