[Mix] Print basic blocks alongside all instructions in BTA debug output
This commit comments all instructions printed during BTA with the names
of parent basic blocks. In order not to complicate the interface of
printing routines with the slot tracker state, printing routines are
factored into BindingTimeAnalysis class, with ModuleSlotTracker as a
state variable there.

As a side effect, value names are now printed more generically, using
proper quoting (via `printLLVMNameWithoutPrefix'), and empty names are
represented as either slot indices or opcode names.