[Mix] Fix the order of debug BTA output
This commit moves debug logging from the sites where WorklistItems are
first put in the queue to where they are popped and executed upon in
BindingTimeAnalysis::fix. For this the WorklistItem is extended with
member variables describing the reason and the `print' function.

The order of output is further stabilized by extending WorklistItem with
the sequential ID field.

Furthermore, in order to cut down the amount of output,
BindingTimeAnalysis class is extended with queue member functions that
transparently drop superseded WorklistItems based on its current and
incoming stage.