[Mix] Expand intrinsic inline
This commit changes the `llvm.mir.ir' expansion from a call to a
generated function to the code of the function itself. This is an
expected way an intrinsic should be expanded, and it avoids creating new
symbols (functions) in the module, which would be especially annoying
for Mix specializers generated by Clang frontend.