[Mix] Extend stage attribute to return parameters and functions
The absense of a return stage makes BTA interprocedural, so an attribute
is required in order to make it global.

Function stage attribute is necessary because both MaxArgStage and
MaxArgStage+1 are valid options for the number of times the function may
be staged. It also works as a flag that indicates that the function is
going to be staged, and BTA may skip functions that are not, potentially
saving lots of compilation time and object space. There are also no
reasonable defaults for "external" functions, given that multi-stage
compilation has potentially infinitely many dynamic stages.