[Mix] Compute object stage as LastStage+1 for non-constant objects
Since the object stage determines at which stages the object is allowed
we need to be stored to, in total `LastStage+2' states are required.
This commit changes `getObjectStage' to return `LastStage+1' if the
object is not a constant at any stage: in this case both loads and
stores are emitted at the last function stage.