[Mix] Include location info in DiagnosticInfoBindingTime
This commit changes the base class and the interface of
DiagnosticInfoBindingTime to include a Function and a
DiagnosticLocation. The function is passed explicitly, and the
DiagnosticLocation is inferred from debug info metadata. Each context
value is now emitted as a separate diagnostic with its own
DiagnosticLocation. This makes the diagnostics much more informative, as
well as simplifies the interface.

Diagnostic tests are updated with debug info and now check that the
correct locations are printed.