[Mix] Access Mix codegen parameters via context table
Move dynamic context, builder, and module parameters used in
StagedIRBuilder to a context table managed by MixContextTable class, and
access it through a dynamic table pointer.

This indirection reduces dependencies on IR values declared in the
parent function outside the current MixFunction, and allows to generate
each MixFunction separately (as opposed to inline), which is required
for static recursion.