[Mix] Move MixContext implementation into a .cpp file
This commit refactors MixContext interface to eliminate usage of
IRBuilder, which removes a template parameter and allows to split the
implementation of MixContextTable and MixContext into a `.cpp` file.

The implementation creates builders by itself, and insert positions are
specified in the call to `MixContextTable::build` or one of the
MixContext constructors.

MixContext now builds all table accesses in the entry block of the mix
function, which simplifies deduplication problem a lot.