[Mix] Fix incoming blocks of static phis
Previously incoming blocks of static phis were derived from predecessors
in the source CFG, but that is not correct if the predecessor is a
dynamic block. In this case incoming blocks must be static basic blocks
of said predecessor. This commit adds a function
`StagedIRBuilder::addIncoming` and calls it for each static basic block
containing the predecessor in Mix.