[Mix] Run stage callbacks on constant values
This fixes `whenStaged` to check if the value waited on is a Constant,
and if so stage it and run the callback right away. Since constants are
not inserted into StagedValues map, it is the only chance for that
callback to ever fire.

This fixes the lack of constant incoming values on dynamic phis.