[Mix] Fix stageStatic for static phis with static operands
Previously `stageStatic` did not stage values of static phis because it
was right for phis with dynamic operands that are already staged, but
that is not correct for phis with static operands that do need to be
staged here. This commit splits `stageStatic` implementation for PHINode
into a separate function and passes operand stage parameter from Mix to
make the correct decision.