[IR] Add GraphTraits specializations for Value def-use graphs
This adds GraphTraits specializations for Value type that provide
GraphTraits interface to operands (forward direction) or uses (inverse
direction) of a given Value. This allows applying graph algorithms to
use-def graphs.

Note that iterating in the forward direction requires dyn_casting each
node to User, but that is pretty much inherent to the underlying
representation of these edges.

The specializations are put in `IR/User.h` to avoid introducing a
cyclical dependency between `IR/Value.h` and `IR/User.h`.